+++ src/main/java/me/hextech/remapped/ProjectionUtils.java
@@ -15,9 +15,9 @@ public class ProjectionUtils {
         Vec3d position = camera.getPos();
         Quaternionf rotation = camera.getRotation();
         Vector3f calculation = rotation.conjugate().transform(position.subtract(destination).toVector3f());
-        Integer fov = (Integer)client.options.getFov().getValue();
+        Integer fov = client.options.getFov().getValue();
         int half = client.getWindow().getScaledHeight() / 2;
         double scale = (double)half / ((double)calculation.z() * Math.tan(Math.toRadians(fov / 2)));
-        return new Vec3d((double)calculation.x() * scale, (double)calculation.y() * scale, (double)calculation.z());
+        return new Vec3d((double)calculation.x() * scale, (double)calculation.y() * scale, calculation.z());
     }
 }
