+++ src/main/java/me/hextech/remapped/Scaffold.java
@@ -65,13 +65,13 @@ extends Module_eSdgMXWuzcxgQVaJFmKZ {
     public void onRender3D(MatrixStack matrixStack, float partialTicks) {
         if (this.render.getValue()) {
             if (this.esp.getValue()) {
-                GL11.glEnable((int)3042);
+                GL11.glEnable(3042);
                 double temp = 0.01;
                 for (double i = 0.0; i < 0.8; i += temp) {
                     HoleSnap.doCircle(matrixStack, ColorUtil.injectAlpha(this.color.getValue(), (int)Math.min((double)(this.color.getValue().getAlpha() * 2) / (0.8 / temp), 255.0)), i, new Vec3d(MathUtil.interpolate(Scaffold.mc.player.lastRenderX, Scaffold.mc.player.getX(), partialTicks), MathUtil.interpolate(Scaffold.mc.player.lastRenderY, Scaffold.mc.player.getY(), partialTicks), MathUtil.interpolate(Scaffold.mc.player.lastRenderZ, Scaffold.mc.player.getZ(), partialTicks)), 5);
                 }
-                RenderSystem.setShaderColor((float)1.0f, (float)1.0f, (float)1.0f, (float)1.0f);
-                GL11.glDisable((int)3042);
+                RenderSystem.setShaderColor(1.0f, 1.0f, 1.0f, 1.0f);
+                GL11.glDisable(3042);
             }
             if (this.pos != null) {
                 Vec3d cur = this.pos.toCenterPos();
